Update build process.
