+++ test.js
@@ -1,32 +1,29 @@
-
 // Unit tests for reactive-property.
-var assert = require("assert");
-
+import assert from "assert";
 // If using from the NPM package, this line would be
 // var Graph = require("graph-data-structure");
-var Graph = require("./index.js");
+import Graph from "./dist/index.modern.js";
 
-var outputGraph = require("graph-diagrams")({
+import graph_diagrams from "graph-diagrams";
 
+const outputGraph = graph_diagrams({
   // If true, writes graph files to ../graph-diagrams for visualization.
   outputGraphs: false,
-  project: "graph-data-structure"
+  project: "graph-data-structure",
 });
 
-function output(graph, name){
+function output(graph, name) {
   outputGraph(graph.serialize(), name);
 }
 
-function withWeight(nodeList, weight){
+function withWeight(nodeList, weight) {
   nodeList.weight = weight;
   return nodeList;
 }
 
-describe("Graph", function() {
-
-  describe("Data structure", function() {
-
-    it("Should add nodes and list them.", function (){
+describe("Graph", function () {
+  describe("Data structure", function () {
+    it("Should add nodes and list them.", function () {
       var graph = Graph();
       graph.addNode("a");
       graph.addNode("b");
@@ -36,14 +33,14 @@ describe("Graph", function() {
       output(graph, "ab-nodes");
     });
 
-    it("Should chain addNode.", function (){
+    it("Should chain addNode.", function () {
       var graph = Graph().addNode("a").addNode("b");
       assert.equal(graph.nodes().length, 2);
       assert(contains(graph.nodes(), "a"));
       assert(contains(graph.nodes(), "b"));
     });
 
-    it("Should remove nodes.", function (){
+    it("Should remove nodes.", function () {
       var graph = Graph();
       graph.addNode("a");
       graph.addNode("b");
@@ -52,7 +49,7 @@ describe("Graph", function() {
       assert.equal(graph.nodes().length, 0);
     });
 
-    it("Should chain removeNode.", function (){
+    it("Should chain removeNode.", function () {
       var graph = Graph()
         .addNode("a")
         .addNode("b")
@@ -61,7 +58,7 @@ describe("Graph", function() {
       assert.equal(graph.nodes().length, 0);
     });
 
-    it("Should add edges and query for adjacent nodes.", function (){
+    it("Should add edges and query for adjacent nodes.", function () {
       var graph = Graph();
       graph.addNode("a");
       graph.addNode("b");
@@ -71,7 +68,7 @@ describe("Graph", function() {
       output(graph, "ab");
     });
 
-    it("Should implicitly add nodes when edges are added.", function (){
+    it("Should implicitly add nodes when edges are added.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       assert.equal(graph.adjacent("a").length, 1);
@@ -81,27 +78,25 @@ describe("Graph", function() {
       assert(contains(graph.nodes(), "b"));
     });
 
-    it("Should chain addEdge.", function (){
+    it("Should chain addEdge.", function () {
       var graph = Graph().addEdge("a", "b");
       assert.equal(graph.adjacent("a").length, 1);
       assert.equal(graph.adjacent("a")[0], "b");
     });
 
-    it("Should remove edges.", function (){
+    it("Should remove edges.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       graph.removeEdge("a", "b");
       assert.equal(graph.adjacent("a").length, 0);
     });
 
-    it("Should chain removeEdge.", function (){
-      var graph = Graph()
-        .addEdge("a", "b")
-        .removeEdge("a", "b");
+    it("Should chain removeEdge.", function () {
+      var graph = Graph().addEdge("a", "b").removeEdge("a", "b");
       assert.equal(graph.adjacent("a").length, 0);
     });
 
-    it("Should not remove nodes when edges are removed.", function (){
+    it("Should not remove nodes when edges are removed.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       graph.removeEdge("a", "b");
@@ -110,21 +105,21 @@ describe("Graph", function() {
       assert(contains(graph.nodes(), "b"));
     });
 
-    it("Should remove outgoing edges when a node is removed.", function (){
+    it("Should remove outgoing edges when a node is removed.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       graph.removeNode("a");
       assert.equal(graph.adjacent("a").length, 0);
     });
 
-    it("Should remove incoming edges when a node is removed.", function (){
+    it("Should remove incoming edges when a node is removed.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       graph.removeNode("b");
       assert.equal(graph.adjacent("a").length, 0);
     });
 
-    it("Should compute indegree.", function (){
+    it("Should compute indegree.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       assert.equal(graph.indegree("a"), 0);
@@ -134,7 +129,7 @@ describe("Graph", function() {
       assert.equal(graph.indegree("b"), 2);
     });
 
-    it("Should compute outdegree.", function (){
+    it("Should compute outdegree.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       assert.equal(graph.outdegree("a"), 1);
@@ -143,19 +138,17 @@ describe("Graph", function() {
       graph.addEdge("a", "c");
       assert.equal(graph.outdegree("a"), 2);
     });
-
   });
 
-  describe("Algorithms", function() {
-
-    it("Should detect cycle.", function() {
+  describe("Algorithms", function () {
+    it("Should detect cycle.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       graph.addEdge("b", "a");
       assert(graph.hasCycle());
     });
 
-    it("Should detect cycle (long).", function() {
+    it("Should detect cycle (long).", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       graph.addEdge("b", "c");
@@ -164,21 +157,20 @@ describe("Graph", function() {
       assert(graph.hasCycle());
     });
 
-    it("Should detect cycle (loop).", function() {
+    it("Should detect cycle (loop).", function () {
       var graph = Graph();
       graph.addEdge("a", "a");
       assert(graph.hasCycle());
     });
 
-    it("Should not detect cycle.", function() {
+    it("Should not detect cycle.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       assert(!graph.hasCycle());
     });
 
     // This example is from Cormen et al. "Introduction to Algorithms" page 550
-    it("Should compute topological sort.", function (){
-
+    it("Should compute topological sort.", function () {
       var graph = Graph();
 
       // Shoes depend on socks.
@@ -207,7 +199,7 @@ describe("Graph", function() {
       output(graph, "getting-dressed");
     });
 
-    it("Should compute topological sort, excluding source nodes.", function (){
+    it("Should compute topological sort, excluding source nodes.", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       graph.addEdge("b", "c");
@@ -218,10 +210,9 @@ describe("Graph", function() {
       output(graph, "abc");
     });
 
-    it("Should compute topological sort tricky case.", function (){
-
-      var graph = Graph();     //      a
-                               //     / \
+    it("Should compute topological sort tricky case.", function () {
+      var graph = Graph(); //      a
+      //     / \
       graph.addEdge("a", "b"); //    b   |
       graph.addEdge("a", "d"); //    |   d
       graph.addEdge("b", "c"); //    c   |
@@ -243,11 +234,8 @@ describe("Graph", function() {
       output(graph, "tricky-case");
     });
 
-    it("Should exclude source nodes with a cycle.", function (){
-      var graph = Graph()
-        .addEdge("a", "b")
-        .addEdge("b", "c")
-        .addEdge("c", "a");
+    it("Should exclude source nodes with a cycle.", function () {
+      var graph = Graph().addEdge("a", "b").addEdge("b", "c").addEdge("c", "a");
       var sorted = graph.topologicalSort(["a"], false);
       assert.equal(sorted.length, 2);
       assert.equal(sorted[0], "b");
@@ -256,9 +244,8 @@ describe("Graph", function() {
       output(graph, "cycle");
     });
 
-    it("Should exclude source nodes with multiple cycles.", function (){
+    it("Should exclude source nodes with multiple cycles.", function () {
       var graph = Graph()
-
         .addEdge("a", "b")
         .addEdge("b", "a")
 
@@ -274,16 +261,15 @@ describe("Graph", function() {
       output(graph, "cycles");
     });
 
-    it("Should error on non-DAG topological sort", function() {
+    it("Should error on non-DAG topological sort", function () {
       var graph = Graph();
       graph.addEdge("a", "b");
       graph.addEdge("b", "a");
       assert.throws(graph.topologicalSort);
     });
 
-    it("Should compute lowest common ancestors.", function (){
+    it("Should compute lowest common ancestors.", function () {
       var graph = Graph()
-
         .addEdge("a", "b")
         .addEdge("b", "d")
         .addEdge("c", "d")
@@ -300,38 +286,35 @@ describe("Graph", function() {
     });
   });
 
-  describe("Edge cases and error handling", function() {
-
-    it("Should return empty array of adjacent nodes for unknown nodes.", function (){
+  describe("Edge cases and error handling", function () {
+    it("Should return empty array of adjacent nodes for unknown nodes.", function () {
       var graph = Graph();
       assert.equal(graph.adjacent("a").length, 0);
       assert.equal(graph.nodes(), 0);
     });
 
-    it("Should do nothing if removing an edge that does not exist.", function (){
-      assert.doesNotThrow(function (){
+    it("Should do nothing if removing an edge that does not exist.", function () {
+      assert.doesNotThrow(function () {
         var graph = Graph();
         graph.removeEdge("a", "b");
       });
     });
 
-    it("Should return indegree of 0 for unknown nodes.", function (){
+    it("Should return indegree of 0 for unknown nodes.", function () {
       var graph = Graph();
       assert.equal(graph.indegree("z"), 0);
     });
 
-    it("Should return outdegree of 0 for unknown nodes.", function (){
+    it("Should return outdegree of 0 for unknown nodes.", function () {
       var graph = Graph();
       assert.equal(graph.outdegree("z"), 0);
     });
-
   });
 
-  describe("Serialization", function() {
-
+  describe("Serialization", function () {
     var serialized;
 
-    function checkSerialized(graph){
+    function checkSerialized(graph) {
       assert.equal(graph.nodes.length, 3);
       assert.equal(graph.links.length, 2);
 
@@ -345,67 +328,61 @@ describe("Graph", function() {
       assert.equal(graph.links[1].target, "c");
     }
 
-    it("Should serialize a graph.", function (){
-      var graph = Graph()
-        .addEdge("a", "b")
-        .addEdge("b", "c");
+    it("Should serialize a graph.", function () {
+      var graph = Graph().addEdge("a", "b").addEdge("b", "c");
       serialized = graph.serialize();
       checkSerialized(serialized);
     });
 
-    it("Should deserialize a graph.", function (){
+    it("Should deserialize a graph.", function () {
       var graph = Graph();
       graph.deserialize(serialized);
       checkSerialized(graph.serialize());
     });
 
-    it("Should chain deserialize a graph.", function (){
+    it("Should chain deserialize a graph.", function () {
       var graph = Graph().deserialize(serialized);
       checkSerialized(graph.serialize());
     });
 
-    it("Should deserialize a graph passed to constructor.", function (){
+    it("Should deserialize a graph passed to constructor.", function () {
       var graph = Graph(serialized);
       checkSerialized(graph.serialize());
     });
   });
 
-  describe("Edge Weights", function() {
-
-    it("Should set and get an edge weight.", function (){
+  describe("Edge Weights", function () {
+    it("Should set and get an edge weight.", function () {
       var graph = Graph().addEdge("a", "b", 5);
       assert.equal(graph.getEdgeWeight("a", "b"), 5);
     });
 
-    it("Should set edge weight via setEdgeWeight.", function (){
-      var graph = Graph()
-        .addEdge("a", "b")
-        .setEdgeWeight("a", "b", 5);
+    it("Should set edge weight via setEdgeWeight.", function () {
+      var graph = Graph().addEdge("a", "b").setEdgeWeight("a", "b", 5);
       assert.equal(graph.getEdgeWeight("a", "b"), 5);
     });
 
-    it("Should return weight of 1 if no weight set.", function (){
+    it("Should return weight of 1 if no weight set.", function () {
       var graph = Graph().addEdge("a", "b");
       assert.equal(graph.getEdgeWeight("a", "b"), 1);
     });
-
   });
 
-  describe("Dijkstra's Shortest Path Algorithm", function (){
-
-    it("Should compute shortest path on a single edge.", function (){
+  describe("Dijkstra's Shortest Path Algorithm", function () {
+    it("Should compute shortest path on a single edge.", function () {
       var graph = Graph().addEdge("a", "b");
       assert.deepEqual(graph.shortestPath("a", "b"), withWeight(["a", "b"], 1));
     });
 
-    it("Should compute shortest path on two edges.", function (){
-      var graph = Graph()
-        .addEdge("a", "b")
-        .addEdge("b", "c");
-      assert.deepEqual(graph.shortestPath("a", "c"), withWeight(["a", "b", "c"], 2));
+    it("Should compute shortest path on two edges.", function () {
+      var graph = Graph().addEdge("a", "b").addEdge("b", "c");
+      assert.deepEqual(
+        graph.shortestPath("a", "c"),
+        withWeight(["a", "b", "c"], 2)
+      );
     });
 
-    it("Should compute shortest path on example from Cormen text (p. 659).", function (){
+    it("Should compute shortest path on example from Cormen text (p. 659).", function () {
       var graph = Graph()
         .addEdge("s", "t", 10)
         .addEdge("s", "y", 5)
@@ -417,57 +394,67 @@ describe("Graph", function() {
         .addEdge("x", "z", 4)
         .addEdge("z", "x", 6);
 
-      assert.deepEqual(graph.shortestPath("s", "z"), withWeight(["s", "y", "z"], 5 + 2));
-      assert.deepEqual(graph.shortestPath("s", "x"), withWeight(["s", "y", "t", "x"], 5 + 3 + 1));
+      assert.deepEqual(
+        graph.shortestPath("s", "z"),
+        withWeight(["s", "y", "z"], 5 + 2)
+      );
+      assert.deepEqual(
+        graph.shortestPath("s", "x"),
+        withWeight(["s", "y", "t", "x"], 5 + 3 + 1)
+      );
     });
 
-    it("Should throw error if source node not in graph.", function (){
+    it("Should throw error if source node not in graph.", function () {
       var graph = Graph().addEdge("b", "c");
       assert.throws(() => graph.shortestPath("a", "c"), /Source node/);
     });
 
-    it("Should throw error if dest node not in graph.", function (){
+    it("Should throw error if dest node not in graph.", function () {
       var graph = Graph().addEdge("b", "c");
       assert.throws(() => graph.shortestPath("b", "g"), /Destination node/);
     });
 
-    it("Should throw error if no path exists.", function (){
-      var graph = Graph()
-        .addEdge("a", "b")
-        .addEdge("d", "e");
+    it("Should throw error if no path exists.", function () {
+      var graph = Graph().addEdge("a", "b").addEdge("d", "e");
       assert.throws(() => graph.shortestPath("a", "e"), /No path/);
     });
 
-    it("Should be robust to disconnected subgraphs.", function (){
-      var graph = Graph()
-        .addEdge("a", "b")
-        .addEdge("b", "c")
-        .addEdge("d", "e");
-      assert.deepEqual(graph.shortestPath("a", "c"), withWeight(["a", "b", "c"], 2));
+    it("Should be robust to disconnected subgraphs.", function () {
+      var graph = Graph().addEdge("a", "b").addEdge("b", "c").addEdge("d", "e");
+      assert.deepEqual(
+        graph.shortestPath("a", "c"),
+        withWeight(["a", "b", "c"], 2)
+      );
     });
   });
 
-  describe("hadEdge", function (){
-    it("Should compute hasEdge.", function (){
+  describe("hadEdge", function () {
+    it("Should compute hasEdge.", function () {
       var graph = Graph().addEdge("a", "b");
       assert.equal(graph.hasEdge("a", "b"), true);
       assert.equal(graph.hasEdge("b", "a"), false);
       assert.equal(graph.hasEdge("c", "a"), false);
     });
-  })
+  });
 });
 
-function contains(arr, item){
-  return arr.filter(function (d){
-    return d === item;
-  }).length > 0;
+function contains(arr, item) {
+  return (
+    arr.filter(function (d) {
+      return d === item;
+    }).length > 0
+  );
 }
 
-function comesBefore(arr, a, b){
+function comesBefore(arr, a, b) {
   var aIndex, bIndex;
-  arr.forEach(function (d, i){
-    if(d === a){ aIndex = i; }
-    if(d === b){ bIndex = i; }
+  arr.forEach(function (d, i) {
+    if (d === a) {
+      aIndex = i;
+    }
+    if (d === b) {
+      bIndex = i;
+    }
   });
   return aIndex < bIndex;
 }
