+++ tsconfig.json
@@ -1,19 +1,20 @@
 {
-    "compilerOptions": {
-        "allowJs": false,
-        "declaration": true,
-        "esModuleInterop": true,
-        "module": "CommonJS",
-        "noImplicitAny": true,
-        "noImplicitReturns": true,
-        "noImplicitThis": true,
-        "strict": true,
-        "target": "ES2016",
-        "strictFunctionTypes": true,
-        "strictNullChecks": true,
-        "downlevelIteration": true,
-        "isolatedModules": true
-    },
-    "file": ["index.ts"],
-    "exclude": ["node_modules", "*.d.ts"]
+  "compilerOptions": {
+    "allowJs": false,
+    "declaration": true,
+    "esModuleInterop": true,
+    "module": "ESNext",
+    "noImplicitAny": true,
+    "noImplicitReturns": true,
+    "noImplicitThis": true,
+    "strict": true,
+    "target": "ESNext",
+    "strictFunctionTypes": true,
+    "strictNullChecks": true,
+    "downlevelIteration": true,
+    "isolatedModules": true,
+    "noEmit": true
+  },
+  "file": ["index.ts"],
+  "exclude": ["node_modules", "*.d.ts"]
 }
